RavenDB input plugin - after review comments - #2
