+++ plugins/inputs/ravendb/testdata/databases_min.json
