+++ plugins/inputs/ravendb/testdata/indexes_full.json
@@ -4,18 +4,22 @@
   "Results": [
     {
       "DatabaseName": "db1",
-      "IndexName": "Product/Rating",
-      "Priority": "Normal",
-      "State": "Normal",
-      "Errors": 0,
-      "TimeSinceLastQueryInSec": 3.4712567,
-      "TimeSinceLastIndexingInSec": 3.4642612,
-      "LockMode": "Unlock",
-      "IsInvalid": true,
-      "Status": "Running",
-      "MappedPerSec": 102.34,
-      "ReducedPerSec": 593.23,
-      "Type": "MapReduce"
+      "Indexes": [
+        {
+          "IndexName": "Product/Rating",
+          "Priority": "Normal",
+          "State": "Normal",
+          "Errors": 0,
+          "TimeSinceLastQueryInSec": 3.4712567,
+          "TimeSinceLastIndexingInSec": 3.4642612,
+          "LockMode": "Unlock",
+          "IsInvalid": true,
+          "Status": "Running",
+          "MappedPerSec": 102.34,
+          "ReducedPerSec": 593.23,
+          "Type": "MapReduce"
+        }
+      ]
     }
   ]
 }
