+++ plugins/inputs/ravendb/testdata/server_full.json
