+++ plugins/inputs/ravendb/testdata/server_min.json
